+++ bdc/bdc/__init__.py
@@ -741,7 +741,8 @@ def process_master_notebook(dest_root, notebook, src_path):
                                            exercises=True,
                                            answers=master['answers'],
                                            encoding_in=master['encoding_in'],
-                                           encoding_out=master['encoding_out'])
+                                           encoding_out=master['encoding_out'],
+                                           verbose=be_verbose)
             move_master_notebooks(master, tempdir)
         except Exception as e:
             error("Failed to process {0}\n    {1}: {2}".format(